upgrade to react-native 0.29.2
