+++ RNTwilioExample/android/app/src/main/java/com/rntwilioexample/MainActivity.java
@@ -1,14 +1,6 @@
 package com.rntwilioexample;
 
 import com.facebook.react.ReactActivity;
-import com.facebook.react.ReactPackage;
-import com.facebook.react.shell.MainReactPackage;
-
-import java.util.Arrays;
-import java.util.List;
-
-import com.lugg.ReactNativeConfig.ReactNativeConfigPackage;
-import com.rogchap.react.modules.twilio.TwilioPackage;
 
 public class MainActivity extends ReactActivity {
 
@@ -21,25 +13,4 @@ public class MainActivity extends ReactActivity {
         return "RNTwilioExample";
     }
 
-    /**
-     * Returns whether dev mode should be enabled.
-     * This enables e.g. the dev menu.
-     */
-    @Override
-    protected boolean getUseDeveloperSupport() {
-        return BuildConfig.DEBUG;
-    }
-
-   /**
-   * A list of packages used by the app. If the app uses additional views
-   * or modules besides the default ones, add more packages here.
-   */
-    @Override
-    protected List<ReactPackage> getPackages() {
-      return Arrays.<ReactPackage>asList(
-        new MainReactPackage(),
-        new ReactNativeConfigPackage(),
-        new TwilioPackage()
-      );
-    }
 }
