+++ RNTwilioExample/package.json
@@ -7,10 +7,9 @@
   },
   "dependencies": {
     "qs": "^6.2.0",
-    "react": "15.1.0",
-    "react-native": "^0.28.0",
-    "react-native-button": "^1.6.0",
-    "react-native-config": "0.0.7",
+    "react": "^15.2.1",
+    "react-native": "^0.29.2",
+    "react-native-config": "fabriziomoscon/react-native-config",
     "react-native-twilio": "fabriziomoscon/react-native-twilio#develop",
     "react-redux": "^4.4.5",
     "redux": "^3.5.2",
